# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "컬컴 백오피스 REST API 문서",
        "title": "Culcom Backoffice API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@culcom.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/customers/check-sms": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "현재 지점의 SMS 연동 상태를 확인합니다",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "SMS 연동 상태 확인",
                "responses": {
                    "200": {
                        "description": "성공",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "연동 안됨",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "인증 실패",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/customers/comment": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "고객의 코멘트를 업데이트합니다",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "고객 코멘트 업데이트",
                "parameters": [
                    {
                        "type": "string",
                        "description": "고객 시퀀스",
                        "name": "customer_seq",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "코멘트",
                        "name": "comment",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "성공",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "잘못된 요청",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "인증 실패",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/customers/increment-call": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "고객의 통화 횟수를 1 증가시킵니다",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "고객 통화 횟수 증가",
                "parameters": [
                    {
                        "type": "string",
                        "description": "고객 시퀀스",
                        "name": "customer_seq",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "성공",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "잘못된 요청",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/customers/reservation": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "새로운 예약 정보를 생성하고 구글 캘린더에 이벤트를 추가합니다",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "예약 정보 생성",
                "parameters": [
                    {
                        "type": "string",
                        "description": "고객 시퀀스",
                        "name": "customer_seq",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "발신자",
                        "name": "caller",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "인터뷰 일시 (2006-01-02T15:04:05)",
                        "name": "interview_date",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "성공",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "잘못된 요청",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "인증 실패",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/customers/sms-senders": {
            "get": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "현재 지점의 SMS 발신번호 목록을 조회합니다",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "SMS 발신번호 목록 조회",
                "responses": {
                    "200": {
                        "description": "성공",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "발신번호 없음",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "인증 실패",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/customers/update-name": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "고객의 이름을 업데이트합니다",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "customers"
                ],
                "summary": "고객 이름 업데이트",
                "parameters": [
                    {
                        "type": "string",
                        "description": "고객 시퀀스",
                        "name": "customer_seq",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "고객 이름",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "성공",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "잘못된 요청",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "서버 오류",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "SessionAuth": {
            "type": "apiKey",
            "name": "user-session",
            "in": "cookie"
        }
    }
}